refactor: centralize domain configuration
